backend/services/ai_service/requirements.txt
fastapi~=0.115.12
uvicorn
python-dotenv~=1.1.0
google-generativeai
weasyprint

asyncio~=3.4.3
motor~=3.7.1
pymongo~=4.13.0
protobuf~=5.29.5
pydantic~=2.11.5
starlette~=0.46.2
python-jose~=3.5.0